socialbook/requirements.txt
Django==5.1.6; python_version >= '3.8'
Django==5.1.6; python_version < '3.8'
django-storages[google]==1.14.4
django-environ==0.12.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
google-cloud-secret-manager==2.22.1 # 2.19.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0 # 2.16.0
google-cloud-documentai==3.1.0 # 2.24.2
# google-resumable-media==2.7.2 # 2.7.0
# google-cloud-aiplatform==1.79.0 # 1.75.0
grpc-google-iam-v1==0.14.0 # 0.13.0
Pillow==11.1.0 # 10.3.0
django-phonenumber-field==8.0.0 # 7.3.0
phonenumbers==8.13.53 # 8.13.33
django-import-export==4.3.4 # 3.3.7
sendgrid==6.11.0
cachetools==5.5.1 # 5.3.3
asgiref==3.8.1
python-http-client==3.3.7
pytz==2024.2 # 2022.7.1
sqlparse==0.5.3 # 0.4.4
django-widget-tweaks==1.5.0
requests==2.32.3 # 2.31.0
testfixtures==8.3.0 # 7.1.0
djangorestframework==3.15.2 # 3.15.1
# django-rest-authtoken==2.1.4
# cryptography==42.0.5
# django-rest-knox==4.2.0
django-cors-headers==4.6.0 # 4.3.1
# djangoconvertvdoctopdf==1.0.1
# python-docx==1.1.2
# reportlab==4.2.5 # 4.2.0
# aspose-words==24.5.0
# Convert2PDF==0.3
# docx2txt==0.8
django-jazzmin==3.0.1
poetry==2.0.1 # 1.8.3
# django-unfold==0.28.1
# fontawesomefree==6.5.1
django-extensions==3.2.3
werkzeug==3.1.3 # 3.0.3
pyopenssl==25.0.0 # 24.1.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0 # 5.3.1
aiohttp==3.11.11
# asyncio==3.4.3
certifi==2024.12.14
# celery==5.5.3
# django-celery-beat==2.7.0
# django-celery-results==2.5.1
# redis==4.6.0
fontawesome-free==5.15.4
pillow-heif==0.22.0
beautifulsoup4==4.13.4
django-otp==1.6.0
qrcode==8.1
openai==1.93.0
django-cron==0.6.0
google-auth==2.43.0